Move versioned/projected maps to utils
